% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adding.R
\name{average}
\alias{average}
\title{Average Function}
\usage{
average(a, b)
}
\arguments{
\item{a}{first number}

\item{b}{second number}
}
\value{

}
\description{
Calculate the average of two numbers
}
\examples{
average(1,2)
}
