% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDP.R
\docType{data}
\name{GDP}
\alias{GDP}
\title{Real GDP of USA}
\format{
295 * 1 dataframe...

\describe{
\item{index}{Date}
\item{column1}{Real GDP}
}
}
\usage{
GDP
}
\description{
Real GDP is the inflation adjusted value of GDP in the United States.
}
\keyword{datasets}
