% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNP.R
\docType{data}
\name{NNP}
\alias{NNP}
\title{NNP of USA}
\format{
50 * 3 dataframe...

\describe{
\item{index}{year}
\item{column1}{GNP}
\item{column2}{Consumption.of.Fixed.Capital}
\item{column3}{NNP}
}
}
\usage{
NNP
}
\description{
NNP of USA
}
\keyword{datasets}
