% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NRoot.R
\name{NRoot}
\alias{NRoot}
\title{NRoot}
\usage{
NRoot(a, b, c)
}
\arguments{
\item{a}{Coefficient of (x^2)}

\item{b}{Coefficient of (x)}

\item{c}{Constant term}
}
\value{

}
\description{
This equation can get the number of real roots of a quadratic function( a(x^2) + bx + c = 0 ).
}
\examples{
NRoot(1,2,1)
}
