% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSquared.R
\name{RSquared}
\alias{RSquared}
\title{RSquared}
\usage{
RSquared(SSE, SST)
}
\arguments{
\item{SSE}{sum of squares of errors(residuals)}

\item{SST}{total sum of squares}
}
\value{

}
\description{
Get the R-Squared of a regression, given SSE and SST.
}
\examples{
RSquared(500,800)
}
