% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sn.R
\name{Sn}
\alias{Sn}
\title{Sn}
\usage{
Sn(a1, n, d)
}
\arguments{
\item{a1}{the first term in the sequence}

\item{n}{the number of the members of the series}

\item{d}{the common difference between terms}
}
\value{

}
\description{
Get the sum of the members of a finite arithmetic sequence.
}
\examples{
Sn(1,5,1)
}
