% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstat.R
\name{tstat}
\alias{tstat}
\title{t-statistic}
\usage{
tstat(b, b0, se)
}
\arguments{
\item{b}{estimate of the parameter}

\item{b0}{the known constant in the null hypothesis}

\item{se}{standard error of the estimator}
}
\value{

}
\description{
Calculate the t-statistic of the data.
}
\examples{
tstat(1.15,1,0.039)
}
