% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VehicleSales.R
\docType{data}
\name{VehicleSales}
\alias{VehicleSales}
\title{Total Vehicle Sales in USA}
\format{
360 * 2 dataframe...

\describe{
\item{column1}{DATE}
\item{column2}{Total Sales}
}
}
\usage{
VehicleSales
}
\description{
Total Vehicle Sales in USA
}
\keyword{datasets}
